Move files to Remote.Razor and update namespaces
May as well do this in the feature branch, since doing it in main would be conflict city
